% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelexamples_Archimedean_copulas.r
\name{ArchimedeanCopulas}
\alias{ArchimedeanCopulas}
\title{Archimedean Copulas}
\usage{
ArchimedeanCopulas()
}
\value{
Returns a list object with elements:
\itemize{
\item \code{onestageparamatric} for the 1-stage parametric estimation
\item \code{twostageparametric} for the 2-stage parametric estimation
\item \code{twostagesemiparametric} for the 2-stage semiparametric estimation
}
These 3 list elements all contain the following items:
\itemize{
\item \code{estimate}: contains all parameter estimates for all models (model 0, model 1, model 2 and model 3)
\item \code{LLH}: contains the maximum likelihood for all models (model 0, model 1, model 2 and model 3)
}
}
\description{
Modelling association through Archimedean Copulas on the Udder Quarter Infection Data Set. For more information see Details.
}
\details{
TO ADD: What is model0, model1, model2, model3 and what is 1stage, 2stageparam, 2stagesemiparam
}
\section{R Code for Model }{

The source R code for this model can found:
\itemize{
\item in the \code{doc/R_Code_ArchimedeanCopulas.R} file in the package installation folder.
\item by accessing the function by calling \code{ArchimedeanCopulas} (without brackets) or \code{getAnywhere("ArchimedeanCopulas")}.
}
}

\examples{
\dontrun{
library(UdderQuarterInfectionData)
data("udderquarterinfection")

ArchimedeanCopulas()
$onestageparametric
$onestageparametric$estimate
model0     model1     model2     model3
lambda1 0.11065835 0.11093770 0.11153123 0.11149235
lambda2 0.11715109 0.11761609 0.11804901 0.11798988
lambda3 0.09548527 0.09489313 0.09520339 0.09526072
lambda4 0.10343006 0.10027196 0.10092973 0.10099590
rho1    1.32062661 1.29720088 1.29925886 1.29854377
rho2    1.26976209 1.26237703 1.26433637 1.26371862
rho3    1.32456636 1.30952760 1.31033903 1.31108332
rho4    1.27967022 1.26625150 1.26907445 1.26965544
beta    0.41840498 0.34405757 0.33992839 0.34008593
theta0          NA 3.18365644 3.05036294 3.04767288
theta1          NA         NA 3.55202588 3.58861334
theta2          NA         NA         NA 3.51261827

$onestageparametric$LLH
model0    model1    model2    model3
-4939.196 -3970.024 -3962.946 -3962.918


$twostageparametric
$twostageparametric$estimate
model1     model2     model3
lambda1 0.11065835 0.11065835 0.11065835
lambda2 0.11715109 0.11715109 0.11715109
lambda3 0.09548527 0.09548527 0.09548527
lambda4 0.10343006 0.10343006 0.10343006
rho1    1.32062661 1.32062661 1.32062661
rho2    1.26976209 1.26976209 1.26976209
rho3    1.32456636 1.32456636 1.32456636
rho4    1.27967022 1.27967022 1.27967022
beta    0.41840498 0.41840498 0.41840498
theta0          NA 2.82505616 2.82128693
theta1          NA 3.29901803 3.36289080
theta2          NA         NA 3.23057648

$twostageparametric$LLH
model1    model2    model3
NA -3965.847 -3965.733


$twostagesemiparametric
$twostagesemiparametric$estimate
model1     model2     model3
lambda1 0.11065835 0.11065835 0.11065835
lambda2 0.11715109 0.11715109 0.11715109
lambda3 0.09548527 0.09548527 0.09548527
lambda4 0.10343006 0.10343006 0.10343006
rho1    1.32062661 1.32062661 1.32062661
rho2    1.26976209 1.26976209 1.26976209
rho3    1.32456636 1.32456636 1.32456636
rho4    1.27967022 1.27967022 1.27967022
beta    0.41840498 0.41840498 0.41840498
theta0  3.22655303 3.11045879 3.10717865
theta1          NA 3.62590946 3.67435249
theta2          NA         NA 3.57471851

$twostagesemiparametric$LLH
model1    model2    model3
-404.3800 -396.6815 -396.6259
}
}
\references{
to add
}
\author{
to add
}
