% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelexamples.R
\name{Gamma_Frailty_Interval_Censoring}
\alias{Gamma_Frailty_Interval_Censoring}
\title{Gamma Frailty Interval Censoring}
\usage{
Gamma_Frailty_Interval_Censoring(print.level = 2)
}
\arguments{
\item{print.level}{Parameter of \code{\link[stats]{nlm}} (default=2): this argument determines the level of printing which is done during the minimization process. The default value of 0 means that no printing occurs, a value of 1 means that initial and final details are printed and a value of 2 means that full tracing information is printed.}
}
\value{
Returns a list with the NLM result in \code{nlm} and the covariance matrix in \code{covmat}.
}
\description{
Application of the Gamma Frailty Interval Censoring Model on the Udder Quarter Infection Data Set. For more information see Details.
}
\details{
ADD MODEL DETAILS HERE
}
\section{R Code for Model }{

The source R code for this model can found:
\itemize{
\item in the \code{doc/Models_R_Code.R} file in the package installation folder.
\item by accessing the function by calling \code{Gamma_Frailty_Interval_Censoring} (without brackets) or \code{getAnywhere("Gamma_Frailty_Interval_Censoring")}.
}
}

\references{
?
}
\author{
?
}
